refactor, code clean up, styles and update device video page
