+++ src/components/Forms/IntakeForm/_intakeForm.scss
@@ -1,10 +1,10 @@
-@import '../../styles/styles';
+@import '../../../styles/styles';
 
 .intake-form {
   box-sizing: border-box;
   width: 400px;
   max-width: 100%;
-  padding: 25px;
+  padding: 0 25px;
   margin: 10px auto 30px;
   &-input {
     margin: $margin-l 0;
@@ -14,5 +14,13 @@
     border-top: none;
     border-right: none;
     border-left: none;
+  }
+  .btn_sizeMin.two {
+    background: $dark;
+    color: $white;
+    &:hover {
+      background: $white;
+      color: $dark;
+    }
   }
 }
