+++ src/components/components.jsx
@@ -3,11 +3,11 @@ export { default as LoginForm } from './LoginForm/LoginForm';
 export { default as LogOutButton } from './LogOutButton/LogOutButton';
 export { default as MobileNav } from './MobileNav/MobileNav';
 export { default as Nav } from './Nav/Nav';
-export { default as IntakeFormBase } from './IntakeForm/IntakeFormBase';
-export { default as IntakeFormFurnace } from './IntakeForm/IntakeFormFurnace';
-export { default as IntakeFormFridge } from './IntakeForm/IntakeFormFridge';
-export { default as IntakeFormUnder_Sink } from './IntakeForm/IntakeFormUnder_Sink';
-export { default as IntakeFormGutters_n_Spouts } from './IntakeForm/IntakeFormGutters_n_Spouts';
+export { default as IntakeFormBase } from './Forms/IntakeForm/IntakeFormBase';
+export { default as IntakeFormFurnace } from './Forms/IntakeForm/IntakeFormFurnace';
+export { default as IntakeFormFridge } from './Forms/IntakeForm/IntakeFormFridge';
+export { default as IntakeFormUnder_Sink } from './Forms/IntakeForm/IntakeFormUnder_Sink';
+export { default as IntakeFormGutters_n_Spouts } from './Forms/IntakeForm/IntakeFormGutters_n_Spouts';
 export { default as SignUpForm } from './SignUpForm/SignUpForm';
 export { default as DevicesListItem } from './DevicesListItem/DevicesListItem';
-
+
