+++ src/pages/DeviceHowToVideos/_deviceHowToVideos.scss
@@ -1,0 +1,5 @@
+.how-to-container {
+  .btn_asLink {
+    margin: 0.25em 0;
+  }
+}
