+++ src/pages/MainIntakePage/FurnaceIntakePage.jsx
@@ -1,18 +1,16 @@
 import { PageLayout } from '../pages';
+import { IntakeFormFurnace } from '../../components/components';
 import '../../styles/_styles.scss';
 
 // CUSTOM COMPONENTS
-import IntakeFormFurnace from '../../components/IntakeForm/IntakeFormFurnace';
-
 export default function FurnaceIntakePage() {
   return (
     <PageLayout>
       <div className="header-three jc-center">Intake Screen</div>
       <div className="">
-        <div className="header-two jc-center">Furnace</div>
+        <div className="header-four jc-center">Furnace</div>
         <IntakeFormFurnace />
-        <div className="row jc-center">
-        </div>
+        <div className="row jc-center"></div>
       </div>
     </PageLayout>
   );
