+++ src/pages/MainIntakePage/Gutters_n_SpoutsPage.jsx
@@ -1,19 +1,18 @@
 import { PageLayout } from '../pages';
 import '../../styles/_styles.scss';
+import { IntakeFormGutters_n_Spouts } from '../../components/components';
 
 // CUSTOM COMPONENTS
-import IntakeFormGutters_n_Spouts from '../../components/IntakeForm/IntakeFormGutters_n_Spouts';
 
-export default function Gutters_n_SpoutsPage (){
+export default function Gutters_n_SpoutsPage() {
   return (
-      <PageLayout>
-        <div className="header-three jc-center">Intake Screen</div>
-        <div className="">
-          <div className="header-two jc-center">Gutters n' Spouts</div>
-          <IntakeFormGutters_n_Spouts />
-          <div className="row jc-center">
-          </div>
-        </div>
-      </PageLayout>
-    );
+    <PageLayout>
+      <div className="header-three jc-center">Intake Screen</div>
+      <div className="">
+        <div className="header-two jc-center">Gutters n' Spouts</div>
+        <IntakeFormGutters_n_Spouts />
+        <div className="row jc-center"></div>
+      </div>
+    </PageLayout>
+  );
 }
